Refactor class mediator converter and fixtures
