+++ cli-synapse/src/test/resources/ballerina/ClassMediatorBasic/functions.bal
@@ -1,4 +1,30 @@
 import ballerina/http;
 
-function greetMediator(http:Response response, string lang) {
+function greetMediator(Context ctx, string lang) {
+}
+
+function respond(Context ctx) returns error? {
+    http:Response response = new;
+    response.setPayload(ctx.payload);
+    foreach [string, string] [name, value] in ctx.headers.entries() {
+        response.setHeader(name, value);
+    }
+    int? statusCode = ctx.statusCode;
+    if statusCode is int {
+        response.statusCode = statusCode;
+    }
+    check (<http:Caller>ctx.caller)->respond(response);
+}
+
+function emitPayload(Context ctx, http:Request request) returns error? {
+    string contentType = request.getContentType();
+    if contentType.startsWith("application/json") {
+        ctx.payload = check request.getJsonPayload();
+    } else if contentType.startsWith("application/xml") || contentType.startsWith("text/xml") {
+        ctx.payload = check request.getXmlPayload();
+    } else if contentType.startsWith("text/") {
+        ctx.payload = check request.getTextPayload();
+    } else {
+        ctx.payload = check request.getBinaryPayload();
+    }
 }
