+++ cli-synapse/src/test/resources/ballerina/ClassMediatorBasic/types.bal
@@ -1,0 +1,12 @@
+import ballerina/http;
+
+public type Variables record {|
+|};
+
+public type Context record {|
+    Variables variables;
+    anydata payload = ();
+    map<string> headers = {};
+    int statusCode?;
+    http:Caller caller?;
+|};
